Simplify screenshot combining & case for no change
There was an issue where "Max Link Length X Days" would be read as
"‘31llvvx A-u yr31 km-days". This is probably because in the first
screenshot this can be where the text starts to fade towards the bottom.
Instead of looking for duplicate characters, just take the first 19
lines from the first screenshot of the all time page (through Combat)
and the last 9 lines from the second screenshot. I also renamed the
variable for clarity. Finally, I added a case for the text summary when
there is no change, instead of just a period preceeded by no comma-
separated updates.