Fix OCR "C]" error and outline sanity checks
Occasionally tesseract is recognizing a '0' character as "C]" due to the
Ingress font. I added a hardcoded replacement to correct this. Further,
sometimes a '0' is read as a '8', but there is no easy way to
differentiate this from a real 8.

One solution is to check if any stats that should only ever increase
have decreased since last recording. This would mean that the last
record incorrectly read a 0 as an 8, and that should be corrected.

Unfortunately, this proposed solution does have issues. It would work
well if the mistaken digit was in the hundred-thousands place, but if
it were in a less-significant place the new value might have surpassed
it. This would not trigger the strictly-increasing check and the
incorrect digit would never be corrected. Further, if multiple 8's are
present in the past value that was found to be invalid, it might be
ambiguous as to which 8(s) was(were) supposed to be a 0.

The real solution would be to learn how to re-train tesseract on the
Ingress font.