Add dependencies for lambda deployment
While normally I wouldn't commit library files and consider this bad
practice, it's common to run into issues preparing a package for use in
AWS Lambda. Often requirements need to be installed on an EC2 instance
running an Amazon Linux AMI for things to run correctly. Also, this
commit includes a ready-to-use instance of tesseract. All of these files
were procured from @Ango's comment on @barbolo's gist:
https://gist.github.com/barbolo/e59aa45ec8e425a26ec4da1086acfbc7#gistcomment-2276042

At this point the files in this repository can be zipped and uploaded
to AWS Lambda to run as-is. The only other AWS requirements are an
API Gateway endpoint to trigger the Lambda, and a DynamoDB table named
"ingress-stats-ocr" in the us-east-1 region. Finally, a Shortcut on iOS
should be used to crop and base64 encode the screenshots and to make an
API call to the Lambda (through API Gateway).