Document iOS Shortcut in repository
An iOS Shortcut is used to crop the screenshots and pass them through an
API call to the Lambda.

Because I cannot export this as a .shortcut file, I'm creating this
markdown file with the iCloud link to download it. Also because you
cannot preview this in the browser (you would need to open the link on
an iOS device and preview it in the Shortcuts app), I'm providing a
textual representation of the shortcut for those browsing this
repository on GitHub.

The shortcut takes the latest three screenshots, crops them to only show
the stats values, base64 encodes them, and makes a POST call to the
Lambda via the API Gateway endpoint, passing along the base64 encoded
images as the JSON request body. Finally it takes the response from the
API call and replaces the \n with actual newlines and displays the text
to the user via Quick Look. To clean up, it deletes the 3 screenshots.